Creados los métodos para traer eventos
